Enhance env. for later use in `02_cmake_fetch_cxxet.bats`

+++ tests/integration/suite/02_cmake_fetch_cxxet.bats
@@ -6,17 +6,24 @@ load "${CUSTOM_BATS_HELPERS_DIRECTORY}/user_log"
 #load "${CUSTOM_BATS_HELPERS_DIRECTORY}/refute_sanitizer_output" # TODO most probably not needed ...
 
 function setup_file() {
-    # TODO set those or not?!
-    #export CXXET_DEFAULT_BLOCK_SIZE=2
-    #export CXXET_VERBOSE=1
-    # TODO guard against non-release preset mode "${CXXET_PRESET}" - see `tests/integration/bats_tests_runner.bash` around line 50
-    export TMP_RESULT_DIR="$(mktemp -d "${TMPDIR:-/tmp}/cxxet.cmake_fetch_cxxet.bats.XXXXXX")"
-}
-
-function setup() {
     if [[ "${CXXET_PRESET:-release}" != 'release' ]]; then
         skip "this should test only 'release' build(s), current preset is '${CXXET_PRESET}'"
     fi
+
+    # TODO set those or not?!
+    #export CXXET_DEFAULT_BLOCK_SIZE=2
+    #export CXXET_VERBOSE=1
+    export TMP_RESULT_DIR="$(mktemp -d "${TMPDIR:-/tmp}/cxxet.cmake_fetch_cxxet.bats.XXXXXX")"
+
+    user_log "# using tmp dir '%s', repository in '%s' and testing its commit '%s'\n" \
+        "${TMP_RESULT_DIR}" "${CXXET_PWD:?}" "${CXXET_CURRENT_COMMIT_HASH:?}"
+    if [[ -n "${CXXET_UNCOMMITED_CHANGES}" ]]; then
+        user_log "\n# !!! BUT BEWARE, THERE ARE UNCOMMITTED CHANGES THAT WON'T BE TESTED !!!\n\n"
+    fi
+}
+
+function setup() {
+    :
 }
 
 function teardown() {
@@ -24,7 +31,7 @@ function teardown() {
 }
 
 function teardown_file() {
-    rm -rf "${TMP_RESULT_DIR}"
+    [[ -z "${TMP_RESULT_DIR}" ]] || rm -rf "${TMP_RESULT_DIR}"
     #user_log "# results from this run are in '%s'\n" "${TMP_RESULT_DIR}"
 }
